requirements.txt
python-dotenv
meilisearch